Moved the credits tests into package, moved tests related to referrals
